IT : Prevent infinite loop when reading files with more than 256 rows [#334]

+++ ATL/AudioData/IO/IT.cs
@@ -320,7 +320,7 @@ namespace ATL.AudioData.IO
         private void readPatterns(BufferedBinaryReader source, IList<uint> patternPointers)
         {
             ushort nbRows;
-            byte rowNum;
+            ushort rowNum;
             byte what;
             byte maskVariable = 0;
             IList<Event> aRow;
